ALGOL68 - Flagged Bubble Sort
I made a flagged bubble sort first because the algorithm for it is so easy, I wanted to make sure any problems I was having was because ALGOL syntax rather than sorting logic or something. I figure we can just include this in out project too, just as an extra thing, since I made it so I could learn ALGOL better.